requirements.txt
# ==========================================================
# Agente Gestor de Proyectos (RAG ClickUp + Chainlit)
# Versión optimizada (GPU o CPU)
# ==========================================================

# 🔹 Núcleo LLM + interfaz
chainlit==2.8.4
python-dotenv==1.2.1
requests==2.32.3
urllib3==2.2.2
certifi==2025.10.5
charset-normalizer==3.4.0
idna==3.10

# 🔹 Calidad y desarrollo
black>=24.0.0
pylint>=3.2.0
mypy>=1.11.0
types-requests
types-setuptools

# ==========================================================
# 🔸 Módulo RAG híbrido (ClickUp → Limpieza → Indexación)
# ==========================================================

# 🔹 Utilidades
pandas==2.2.3
numpy==1.26.4
tqdm==4.66.5
rich==13.9.2
pyyaml>=6.0.2

# 🔹 Procesamiento y splitting de texto
nltk==3.9.1
langchain-text-splitters==0.3.6
regex>=2024.5.15

# 🔹 Indexación y búsqueda híbrida
chromadb==0.5.5
rank-bm25==0.2.2
sentence-transformers==3.1.1
scikit-learn==1.5.2

# ==========================================================
# 🔸 Modelos neuronales (Hugging Face)
# ==========================================================
transformers==4.45.2
accelerate>=0.33.0
safetensors>=0.4.3
bitsandbytes==0.43.3      # para 4bit/8bit sin CUDA SDK completo
torch==2.4.0
datasets>=3.0.1

# ==========================================================
# 🔸 API OpenAI (opcional, para pruebas)
# ==========================================================
openai>=1.0.0,<2.0.0

# ==========================================================
# 🔸 API/Servidor y tests
# ==========================================================
fastapi>=0.116.1,<0.117.0
uvicorn[standard]>=0.35.0
aiofiles>=23.2.1
pytest>=8.3.0
pytest-asyncio>=0.24.0

# ==========================================================
# End of requirements.txt
# ==========================================================
